+++ src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS annotations(
    id BIGINT PRIMARY KEY AUTO_INCREMENT,
    description VARCHAR(255) NOT NULL,
-   value DOUBLE NOT NULL,
+   value DECIMAL(19,2) NOT NULL,
    created_at DATE NOT NULL,
    type VARCHAR(255) NOT NULL
 );
